+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/jss/compare/v1.35.0...HEAD)
+[Unreleased](https://github.com/jshwi/jss/compare/v1.36.0...HEAD)
+------------------------------------------------------------------------
+
+[1.36.0](https://github.com/jshwi/jss/releases/tag/v1.36.0) - 2023-12-05
 ------------------------------------------------------------------------
 ### Changed
 - only add users that are authorized to sitemap